+++ bazzar_base/src/com/bazzar/base/service/CartService.java
@@ -4,7 +4,7 @@ import com.bazzar.base.domain.order.Cart;
 
 public interface CartService {
 
-	public int create ( Cart cart );
+	public Long create ( Cart cart );
 	public void delete ( Cart cart );
 	public void delete ( Long cartId );
 	public void edit ( Cart cart );
