+++ bazzar_base/src/com/bazzar/base/service/OrderService.java
@@ -9,7 +9,7 @@ public interface OrderService {
 	public List <Order> getOrders ();
 	public List <Order> getOrdersForCustomer ( Long customerId );
 	public Order getOrder ( Long orderId );
-	public int createOrder ( Order order );
+	public Long createOrder ( Order order );
 	public void editOrder ( Order order );
 	public Order getOrderByInvoice ( String invoiceNumber );
 	public void delete ( Long id );
